Add commented configuration for postfix_exporter example (#25)
* Add commented configuration for postfix_exporter

Provide a commented-out example for configuring postfix_exporter in `stack.yml`. This serves as a reference for potential future use, including network and volume bindings.

Signed-off-by: Jochen Issing <c.333+git@nesono.com>

* Add a TODO comment for postfix_exporter migration

A TODO comment was added to clarify the need for uncommenting the postfix_exporter section after migrating to Docker Compose. This avoids issues with Docker name resolution during the transition.

Signed-off-by: Jochen Issing <c.333+git@nesono.com>

---------

Signed-off-by: Jochen Issing <c.333+git@nesono.com>
Co-authored-by: Jochen Issing <c.333+git@nesono.com>