Add spawn protection and update server operators list (#28)
* Add spawn protection and new OP to Minecraft server config

Introduced a `SPAWN_PROTECTION` environment variable with a value of 0 to the server configuration. Additionally, added `PlageGeist0903` to the list of server operators (OPS).

Signed-off-by: Jochen Issing <c.333+git@nesono.com>

* Remove extra blank line in docker-compose.yaml

    This change removes an unnecessary blank line from the docker-compose.yaml file. It ensures the file adheres to cleaner formatting and improves readability.

Signed-off-by: Jochen Issing <c.333+git@nesono.com>

* Update MOTD in docker-compose.yaml

The Message of the Day (MOTD) has been updated to "Hello hello hello!" from the previous value. This change reflects a more standardized or welcoming message for users.

Signed-off-by: Jochen Issing <c.333+git@nesono.com>

---------

Signed-off-by: Jochen Issing <c.333+git@nesono.com>
Co-authored-by: Jochen Issing <c.333+git@nesono.com>